+++ src/braket/circuits/ascii_circuit_diagram.py
@@ -86,7 +86,10 @@ class AsciiCircuitDiagram(CircuitDiagram):
 
         # A list of parameters in the circuit to the currently assigned values.
         if circuit.parameters:
-            lines.append(f"\nUnassigned parameters: {circuit.parameters}.")
+            lines.append(
+                "\nUnassigned parameters: "
+                f"{sorted(circuit.parameters, key=lambda param: param.name)}."
+            )
 
         return "\n".join(lines)