+++ src/braket/circuits/noise_model/__init__.py
@@ -1,0 +1,30 @@
+# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+#     http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+
+from braket.circuits.noise_model.circuit_instruction_criteria import (  # noqa: F401
+    CircuitInstructionCriteria,
+)
+from braket.circuits.noise_model.criteria import (  # noqa: F401
+    Criteria,
+    CriteriaKey,
+    CriteriaKeyResult,
+)
+from braket.circuits.noise_model.gate_criteria import GateCriteria  # noqa: F401
+from braket.circuits.noise_model.initialization_criteria import InitializationCriteria  # noqa: F401
+from braket.circuits.noise_model.noise_model import NoiseModel, NoiseModelInstruction  # noqa: F401
+from braket.circuits.noise_model.observable_criteria import ObservableCriteria  # noqa: F401
+from braket.circuits.noise_model.qubit_initialization_criteria import (  # noqa: F401
+    QubitInitializationCriteria,
+)
+from braket.circuits.noise_model.result_type_criteria import ResultTypeCriteria  # noqa: F401
+from braket.circuits.noise_model.unitary_gate_criteria import UnitaryGateCriteria  # noqa: F401